+++ src/main/java/whu/edu/moniData/BuQuan/buquan2.java
@@ -152,7 +152,7 @@ public class buquan2 {
     }
     private static PathPointData predictNextMixed(long keyInPointMap,String timestamp){
         PathPointData pdInPointMap=pointMap.get(keyInPointMap);
-        Pair<ConcurrentLinkedDeque<Float>,Float> a1=predictSpeedWindow(pdInPointMap);//速度窗口、预测的速度
+        Pair<ConcurrentLinkedDeque<Double>,Double> a1=predictSpeedWindow(pdInPointMap);//速度窗口、预测的速度
         double[] a2=predictNewMileage(pdInPointMap,a1.getValue());//新里程、驶过的距离
         if(a2[0]<1016020||a2[1]>1037954){
             pointMap.remove(keyInPointMap);tempMap.remove(keyInPointMap);
@@ -166,7 +166,7 @@ public class buquan2 {
 
         double carangle=calculateBearing(a3.getValue()[1],a3.getValue()[0],pdInPointMap.getLatitude(),pdInPointMap.getLongitude());
         pdInPointMap.setCarAngle(carangle);
-        pdInPointMap.setMileage((int) (a2[0]));
+        pdInPointMap.setMileage(a2[0]);
         pdInPointMap.setSpeed(a1.getValue());
 //        pdInPointMap.setTimeStamp(pathTimeStamp);//未接收到，不更新
         pdInPointMap.setLatitude(a3.getValue()[1]);
@@ -186,7 +186,7 @@ public class buquan2 {
     }
 
 
-    public static double[] predictNewMileage(PathPointData data,float speed){
+    public static double[] predictNewMileage(PathPointData data,double speed){
 
         double[]d={0,0};
         d[1] = myTools.calculateDistance(speed, 200);
