+++ src/main/java/whu/edu/moniData/FiberGratingData6AllData.java
@@ -783,7 +783,7 @@ public class FiberGratingData6AllData {
 
     private static double[] generateInitialJingWei(long tpointno, int direc) {
         TrafficEventUtils.MileageConverter converter = (direc == 1) ? mileageConverter1 : mileageConverter2;
-        return converter.findCoordinate((int) tpointno).getLnglat();
+        return converter.findCoordinate(tpointno).getLnglat();
     }
 
     private static void closeProducers() {
