+++ src/main/java/whu/edu/moniData/Utils/TrafficEventUtils.java
@@ -23,30 +23,30 @@ public class TrafficEventUtils {
     @Getter
     @Setter
     public static class TrafficEvent implements Serializable{
-        private Integer eventId; // 事件id（唯一性）
-        private String timeStamp; // 事件上报时间戳
-        private Integer eventType; // 事件类型
-        private String startStake; // 事件起始桩号
-        private String endStake; // 事件结束桩号
-        private Integer startMileage; // 事件范围起始里程
-        private Integer endMileage; // 事件范围截止里程
-        private Double startLongitude; // 事件范围起始经度
-        private Double startLatitude; // 事件范围起始纬度
-        private Double endLongitude; // 事件范围截止经度
-        private Double endLatitude; // 事件范围截止纬度
-        private String laneNo; // 事件或施工所在车道号
-        private Integer direction; // 事件发生区域的行车方向
-        private JSONArray carList; // 事件发生车辆集合
-        private String eventLevel; // 事件等级
-        private String eventDes; // 事件概要描述
-        private String eventReason; // 事件原因
-        private String eventPicPath; // 事件现场图片路径，多张逗号分隔
-        private String eventVideoPath; // 事件现场视频路径
-//        private String eventSource; // 事件源
-        private String sourceRemark; // 来源备注
-        private String waySectionId; // 路段id
-        private String waySectionName; // 路段名称
-        private Boolean manualAudit; // 事件是否需要人工审核校验
+        private Integer eventId;          // 事件id（唯一性）
+        private String timeStamp;         // 事件上报时间戳
+        private Integer eventType;        // 事件类型
+        private String startStake;        // 事件起始桩号
+        private String endStake;          // 事件结束桩号
+        private Integer startMileage;     // 事件范围起始里程
+        private Integer endMileage;       // 事件范围截止里程
+        private Double startLongitude;    // 事件范围起始经度
+        private Double startLatitude;     // 事件范围起始纬度
+        private Double endLongitude;      // 事件范围截止经度
+        private Double endLatitude;       // 事件范围截止纬度
+        private String laneNo;            // 事件或施工所在车道号
+        private Integer direction;        // 事件发生区域的行车方向
+        private JSONArray carList;        // 事件发生车辆集合
+        private String eventLevel;        // 事件等级
+        private String eventDes;          // 事件概要描述
+        private String eventReason;       // 事件原因
+        private String eventPicPath;      // 事件现场图片路径，多张逗号分隔
+        private String eventVideoPath;    // 事件现场视频路径
+//        private String eventSource;     // 事件源
+        private String sourceRemark;      // 来源备注
+        private String waySectionId;      // 路段id
+        private String waySectionName;    // 路段名称
+        private Boolean manualAudit;      // 事件是否需要人工审核校验
         private Integer constructionVehicles; // 施工类事件，施工车数量
         private Integer constructionPerson; // 施工类事件，施工人员数量
 
@@ -91,7 +91,7 @@ public class TrafficEventUtils {
         private long carId;
         private String plateNo;
         private Integer plateColor;
-        private float speedSum;
+        private double speedSum;
         private int direction;
         private int pointSum;
         private Integer originalType = null;
@@ -327,8 +327,8 @@ public class TrafficEventUtils {
             }
         }
 
-        private int stakeToMileage(String stakeId) {
-            return Integer.parseInt(stakeId.split("\\+")[0].substring(1)) * 1000 + Integer.parseInt(stakeId.split("\\+")[1]);
+        private double stakeToMileage(String stakeId) {
+            return Double.parseDouble(stakeId.split("\\+")[0].substring(1)) * 1000 + Double.parseDouble(stakeId.split("\\+")[1]);
         }
     }
 
@@ -338,8 +338,8 @@ public class TrafficEventUtils {
     @Getter
     @Setter
     public static class MileageConverter implements Serializable {
-        private Map<Integer, StakeInfo> stakeInfoMap;
-        private List<Integer> mileageList;
+        private Map<Double, StakeInfo> stakeInfoMap;
+        private List<Double> mileageList;
 
         public MileageConverter(String resourceName) throws IOException {
             try (InputStream is = getClass().getClassLoader().getResourceAsStream(resourceName)) {
@@ -348,11 +348,11 @@ public class TrafficEventUtils {
                 }
                 this.stakeInfoMap = loadCheckpointsFromJSON(is);
                 this.mileageList = new ArrayList<>(stakeInfoMap.keySet());
-                this.mileageList.sort(Integer::compare);
+                this.mileageList.sort(Double::compare);
             }
         }
 
-        private Map<Integer, StakeInfo> loadCheckpointsFromJSON(InputStream is) throws IOException {
+        private Map<Double, StakeInfo> loadCheckpointsFromJSON(InputStream is) throws IOException {
             ByteArrayOutputStream result = new ByteArrayOutputStream();
             byte[] buffer = new byte[1024];
             int length;
@@ -362,10 +362,10 @@ public class TrafficEventUtils {
             String json = new String(result.toByteArray(), StandardCharsets.UTF_8);
             List<StakeInfo> list = JSON.parseObject(json, new TypeReference<List<StakeInfo>>() {});
 
-            Map<Integer, StakeInfo> map = new HashMap<>();
+            Map<Double, StakeInfo> map = new HashMap<>();
             for (StakeInfo info : list) {
                 try {
-                    int mileage = stakeToMileage(info.getStake());
+                    double mileage = stakeToMileage(info.getStake());
                     map.put(mileage, info);
                 } catch (Exception e) {
                     System.err.println("桩号解析失败: " + info.getStake());
@@ -374,14 +374,14 @@ public class TrafficEventUtils {
             return map;
         }
 
-        private List<Integer> loadSortedList(Map<Integer, StakeInfo> stakeInfoMap) {
-            List<Integer> mileageList = new ArrayList<>();
+        private List<Double> loadSortedList(Map<Double, StakeInfo> stakeInfoMap) {
+            List<Double> mileageList = new ArrayList<>();
             mileageList.addAll(stakeInfoMap.keySet());
-            mileageList.sort(Integer::compare);
+            mileageList.sort(Double::compare);
             return mileageList;
         }
 
-        public StakeInfo findCoordinate(int targetMileage) {
+        public StakeInfo findCoordinate(double targetMileage) {
             if(targetMileage < mileageList.get(0))
                 return stakeInfoMap.get(mileageList.get(0));
             else if(targetMileage > mileageList.get(mileageList.size() - 1))
@@ -400,8 +400,8 @@ public class TrafficEventUtils {
             }
         }
 
-        private int stakeToMileage(String stakeId) {
-            return Integer.parseInt(stakeId.split("\\+")[0].substring(1)) * 1000 + Integer.parseInt(stakeId.split("\\+")[1]);
+        private double stakeToMileage(String stakeId) {
+            return Double.parseDouble(stakeId.split("\\+")[0].substring(1)) * 1000 + Double.parseDouble(stakeId.split("\\+")[1]);
         }
 
     }
